requirements.txt
# ================================
# Core Scientific Computing
# ================================
numpy              # Numerical operations and array handling
scipy              # Signal processing and scientific utilities
tqdm               # Progress bars for long-running processes

# ================================
# Audio Processing
# ================================
librosa            # Audio loading, feature extraction (mel spectrograms, etc.)
soundfile          # Reading/writing WAV audio files

# ================================
# Image & Video Processing
# ================================
opencv-python      # Computer vision operations (image/video frame handling)
opencv-contrib-python  # Extra OpenCV modules used by Wav2Lip

# ================================
# Deep Learning (PyTorch)
# ================================
torch              # Neural network framework (Wav2Lip model runs on this)
torchvision        # Image transforms and utilities for PyTorch

# ================================
# Wav2Lip Model Dependencies
# ================================
numba              # Speeds up numerical functions (used by librosa/Wav2Lip)
face-alignment     # Face landmark detection used by Wav2Lip

# ================================
# API & Environment Configuration
# ================================
requests           # HTTP requests for ElevenLabs API calls
python-dotenv      # Loads environment variables from .env file
pydub              # Audio file manipulation (mp3 to wav conversion)